refactor(core): extract distance functions to core-lib crate
Move all distance computation logic to rapid-fuzzy-core to eliminate
duplication between napi and wasm binding crates. Both bindings are
now thin wrappers that delegate to shared pure-Rust implementations.

Closes #505
