+++ crates/core-lib/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod distance;
 pub mod search;
